[BE][CI] Rename .jenkins to .ci (#82753)
### Description
We've wanted to do this for a looong time. Let's eradicate confusion about what the .jenkins folder is and try to host our scripts in a more generic .ci folder.

Most changes are a find and replace EXCEPT
- deleted dirty.sh in both caffe2 and pytorch since it was not used anywhere
- modified CONTRIBUTING.md to be more updated

The next steps:
- rename jenkins user to pytorch-ci or something
- move scripts from .circleci to here (e.g., docker)

### Issue
We used to have an issue for this but it is so long ago that I could not find it.

### Testing
CI should not break.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82753
Approved by: https://github.com/malfet, https://github.com/seemethere, https://github.com/huydhn